Add projectile support to persp-harpoon configuration
Add a new function persp-harpoon-configure-for-projectile that
configures persp-harpoon to integrate with projectile. This includes
adding hooks for projectile project switches to update the harpoon
buffers accordingly. The existing
persp-harpoon-configure-for-perspective function was also updated to
support unsetting hooks and variables. The README was updated to
describe the new projectile support and clarify configuration
options. This extends persp-harpoon's compatibility to projectile in
addition to perspective.el.